Move the tests for WRITING_JOB_ABORTED
